// tina/__generated__/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  /** References another document, used as a foreign key */
  Reference: { input: any; output: any };
  JSON: { input: any; output: any };
};

export type SystemInfo = {
  __typename?: "SystemInfo";
  filename: Scalars["String"]["output"];
  title?: Maybe<Scalars["String"]["output"]>;
  basename: Scalars["String"]["output"];
  breadcrumbs: Array<Scalars["String"]["output"]>;
  path: Scalars["String"]["output"];
  relativePath: Scalars["String"]["output"];
  extension: Scalars["String"]["output"];
  template: Scalars["String"]["output"];
  collection: Collection;
};
// ...
export type Node = {
  id: Scalars["ID"]["output"];
};

export type Document = {
  id: Scalars["ID"]["output"];
  _sys?: Maybe<SystemInfo>;
  _values: Scalars["JSON"]["output"];
};

/** A relay-compliant pagination connection */
export type Connection = {
  totalCount: Scalars["Float"]["output"];
  pageInfo: PageInfo;
};

export type Query = {
  __typename?: "Query";
  getOptimizedQuery?: Maybe<Scalars["String"]["output"]>;
  collection: Collection;
  collections: Array<Collection>;
  node: Node;
  document: DocumentNode;
  globals: Globals;
  globalsConnection: GlobalsConnection;
  page: Page;
  pageConnection: PageConnection;
};
// ...
export type Collection = {
  __typename?: "Collection";
  name: Scalars["String"]["output"];
  slug: Scalars["String"]["output"];
  label?: Maybe<Scalars["String"]["output"]>;
  path: Scalars["String"]["output"];
  format?: Maybe<Scalars["String"]["output"]>;
  matches?: Maybe<Scalars["String"]["output"]>;
  templates?: Maybe<Array<Maybe<Scalars["JSON"]["output"]>>>;
  fields?: Maybe<Array<Maybe<Scalars["JSON"]["output"]>>>;
  documents: DocumentConnection;
};

export type CollectionDocumentsArgs = {
  before?: InputMaybe<Scalars["String"]["input"]>;
  after?: InputMaybe<Scalars["String"]["input"]>;
  first?: InputMaybe<Scalars["Float"]["input"]>;
  last?: InputMaybe<Scalars["Float"]["input"]>;
  sort?: InputMaybe<Scalars["String"]["input"]>;
  filter?: InputMaybe<DocumentFilter>;
  folder?: InputMaybe<Scalars["String"]["input"]>;
};

export type DocumentNode = Globals | Page | Folder;
// ...
export type Globals = Node &
  Document & {
    __typename?: "Globals";
    navigation?: Maybe<GlobalsNavigation>;
    footer?: Maybe<GlobalsFooter>;
    id: Scalars["ID"]["output"];
    _sys: SystemInfo;
    _values: Scalars["JSON"]["output"];
  };
// ...
export type GlobalsConnection = Connection & {
  __typename?: "GlobalsConnection";
  pageInfo: PageInfo;
  totalCount: Scalars["Float"]["output"];
  edges?: Maybe<Array<Maybe<GlobalsConnectionEdges>>>;
};

export type PageBlocksHero = {
  __typename?: "PageBlocksHero";
  image?: Maybe<Scalars["String"]["output"]>;
  heading?: Maybe<Scalars["String"]["output"]>;
};

export type PageBlocksValuesQuality1 = {
  __typename?: "PageBlocksValuesQuality1";
  title?: Maybe<Scalars["String"]["output"]>;
  description?: Maybe<Scalars["JSON"]["output"]>;
};

export type PageBlocksValuesQuality2 = {
  __typename?: "PageBlocksValuesQuality2";
  title?: Maybe<Scalars["String"]["output"]>;
  description?: Maybe<Scalars["JSON"]["output"]>;
};

export type PageBlocksValuesQuality3 = {
  __typename?: "PageBlocksValuesQuality3";
  title?: Maybe<Scalars["String"]["output"]>;
  description?: Maybe<Scalars["JSON"]["output"]>;
};

export type PageBlocksValues = {
  __typename?: "PageBlocksValues";
  backgroundImage?: Maybe<Scalars["String"]["output"]>;
  title?: Maybe<Scalars["String"]["output"]>;
  quality1?: Maybe<PageBlocksValuesQuality1>;
  quality2?: Maybe<PageBlocksValuesQuality2>;
  quality3?: Maybe<PageBlocksValuesQuality3>;
};

export type PageBlocksLocation = {
  __typename?: "PageBlocksLocation";
  content?: Maybe<Scalars["JSON"]["output"]>;
};

export type PageBlocksMachineCarouselMachineCard = {
  __typename?: "PageBlocksMachineCarouselMachineCard";
  image?: Maybe<Scalars["String"]["output"]>;
  name?: Maybe<Scalars["String"]["output"]>;
  description?: Maybe<Scalars["JSON"]["output"]>;
  type?: Maybe<Scalars["String"]["output"]>;
};

export type PageBlocksMachineCarousel = {
  __typename?: "PageBlocksMachineCarousel";
  sectionTitle?: Maybe<Scalars["String"]["output"]>;
  buttonLabel?: Maybe<Scalars["String"]["output"]>;
  machineCard?: Maybe<Array<Maybe<PageBlocksMachineCarouselMachineCard>>>;
};

export type PageBlocks =
  | PageBlocksHero
  | PageBlocksValues
  | PageBlocksLocation
  | PageBlocksMachineCarousel;

export type Page = Node &
  Document & {
    __typename?: "Page";
    blocks?: Maybe<Array<Maybe<PageBlocks>>>;
    id: Scalars["ID"]["output"];
    _sys: SystemInfo;
    _values: Scalars["JSON"]["output"];
  };
// ...
export type PageConnection = Connection & {
  __typename?: "PageConnection";
  pageInfo: PageInfo;
  totalCount: Scalars["Float"]["output"];
  edges?: Maybe<Array<Maybe<PageConnectionEdges>>>;
};
// ...
export type PageQuery = {
  __typename?: "Query";
  page: {
    __typename: "Page";
    id: string;
    _sys: {
      __typename?: "SystemInfo";
      filename: string;
      basename: string;
      breadcrumbs: Array<string>;
      path: string;
      relativePath: string;
      extension: string;
    };
    blocks?: Array<
      | {
          __typename: "PageBlocksHero";
          image?: string | null;
          heading?: string | null;
        }
      | {
          __typename: "PageBlocksValues";
          backgroundImage?: string | null;
          title?: string | null;
          quality1?: {
            __typename: "PageBlocksValuesQuality1";
            title?: string | null;
            description?: any | null;
          } | null;
          quality2?: {
            __typename: "PageBlocksValuesQuality2";
            title?: string | null;
            description?: any | null;
          } | null;
          quality3?: {
            __typename: "PageBlocksValuesQuality3";
            title?: string | null;
            description?: any | null;
          } | null;
        }
      | { __typename: "PageBlocksLocation"; content?: any | null }
      | {
          __typename: "PageBlocksMachineCarousel";
          sectionTitle?: string | null;
          buttonLabel?: string | null;
          machineCard?: Array<{
            __typename: "PageBlocksMachineCarouselMachineCard";
            image?: string | null;
            name?: string | null;
            description?: any | null;
            type?: string | null;
          } | null> | null;
        }
      | null
    > | null;
  };
};